Title: 특정 조건 세 정수 비교, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240806/특정 조건 세 정수 비교/specific-comparison-of-three-natural-numbers.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-    double a, b, c;
+    int a, b, c;
 
     cin >> a >> b >> c;
